Deduplicate interrupt status enum
