+++ esp-hal/src/gpio/interrupt.rs
@@ -62,9 +62,9 @@ use procmacros::ram;
 use strum::EnumCount;
 
 use crate::{
-    gpio::{AnyPin, GpioBank, InputPin, InterruptStatusRegisterAccess, set_int_enable},
+    gpio::{AnyPin, GpioBank, InputPin, set_int_enable},
     interrupt::{self, DEFAULT_INTERRUPT_HANDLER, Priority},
-    peripherals::Interrupt,
+    peripherals::{GPIO, Interrupt},
     sync::RawMutex,
 };
 
@@ -200,6 +200,35 @@ pub(super) extern "C" fn user_gpio_interrupt_handler() {
             handle_async_pins(bank, async_pins, intrs);
         }
     });
+}
+
+#[derive(Clone, Copy)]
+pub(crate) enum InterruptStatusRegisterAccess {
+    Bank0,
+    #[cfg(gpio_has_bank_1)]
+    Bank1,
+}
+
+impl InterruptStatusRegisterAccess {
+    pub(crate) fn interrupt_status_read(self) -> u32 {
+        cfg_if::cfg_if! {
+            if #[cfg(esp32)] {
+                match self {
+                    Self::Bank0 => GPIO::regs().status().read().bits(),
+                    Self::Bank1 => GPIO::regs().status1().read().bits(),
+                }
+            } else if #[cfg(any(esp32c2, esp32c3, esp32c6, esp32h2))] {
+                GPIO::regs().pcpu_int().read().bits()
+            } else if #[cfg(any(esp32s2, esp32s3))] {
+                // Whilst the S3 is a dual core chip, it shares the enable registers between
+                // cores so treat it as a single core device
+                match self {
+                    Self::Bank0 => GPIO::regs().pcpu_int().read().bits(),
+                    Self::Bank1 => GPIO::regs().pcpu_int1().read().bits(),
+                }
+            }
+        }
+    }
 }
 
 fn interrupt_status() -> [(GpioBank, u32); GpioBank::COUNT] {
