+++ esp-hal/src/soc/esp32/gpio.rs
@@ -42,11 +42,7 @@
 
 use core::mem::transmute;
 
-use crate::{
-    pac::io_mux,
-    peripherals::{GPIO, IO_MUX},
-    system::Cpu,
-};
+use crate::{pac::io_mux, peripherals::IO_MUX, system::Cpu};
 
 pub(crate) fn io_mux_reg(gpio_num: u8) -> &'static io_mux::GPIO0 {
     let iomux = IO_MUX::regs();
@@ -748,19 +744,4 @@ touch! {
     // ---
     (8, 33, sar_touch_out5, sar_touch_thres5, false)
     (9, 32, sar_touch_out5, sar_touch_thres5, false)
-}
-
-#[derive(Clone, Copy)]
-pub(crate) enum InterruptStatusRegisterAccess {
-    Bank0,
-    Bank1,
-}
-
-impl InterruptStatusRegisterAccess {
-    pub(crate) fn interrupt_status_read(self) -> u32 {
-        match self {
-            Self::Bank0 => GPIO::regs().status().read().bits(),
-            Self::Bank1 => GPIO::regs().status1().read().bits(),
-        }
-    }
 }
