+++ esp-hal/src/soc/esp32s3/gpio.rs
@@ -36,10 +36,7 @@
 //! registers for both the `PRO CPU` and `APP CPU`. The implementation uses the
 //! `gpio` peripheral to access the appropriate registers.
 
-use crate::{
-    pac::io_mux,
-    peripherals::{GPIO, IO_MUX},
-};
+use crate::{pac::io_mux, peripherals::IO_MUX};
 
 pub(crate) fn io_mux_reg(gpio_num: u8) -> &'static io_mux::GPIO {
     IO_MUX::regs().gpio(gpio_num as usize)
@@ -487,23 +484,6 @@ rtcio_analog! {
     (19, rtc_pad19(),    pad19      )
     (20, rtc_pad20(),    pad20      )
     (21, rtc_pad21(),    pad21      )
-}
-
-// Whilst the S3 is a dual core chip, it shares the enable registers between
-// cores so treat it as a single core device
-#[derive(Clone, Copy)]
-pub(crate) enum InterruptStatusRegisterAccess {
-    Bank0,
-    Bank1,
-}
-
-impl InterruptStatusRegisterAccess {
-    pub(crate) fn interrupt_status_read(self) -> u32 {
-        match self {
-            Self::Bank0 => GPIO::regs().pcpu_int().read().bits(),
-            Self::Bank1 => GPIO::regs().pcpu_int1().read().bits(),
-        }
-    }
 }
 
 fn enable_iomux_clk_gate() {
